+++ api/v1/controllers/class/active.js
@@ -1,5 +1,5 @@
 const { httpPermCheck } = require("@modules/middleware/permission-check");
-const { isClassActive } = require("@modules/class/class");
+const { isClassActive } = require("@services/class-service");
 const { classInformation } = require("@modules/class/classroom");
 const ForbiddenError = require("@errors/forbidden-error");
 const { isAuthenticated } = require("@modules/middleware/authentication");
