+++ api/v1/controllers/class/polls/end.js
@@ -2,7 +2,7 @@ const { hasClassPermission } = require("@modules/middleware/permission-check");
 const { isAuthenticated } = require("@modules/middleware/authentication");
 const { parseJson } = require("@modules/middleware/parse-json");
 const { CLASS_PERMISSIONS } = require("@modules/permissions");
-const { updatePoll } = require("@modules/polls");
+const { updatePoll } = require("@services/poll-service");
 
 module.exports = (router) => {
     /**
