+++ api/v1/controllers/class/polls/response.js
@@ -1,4 +1,4 @@
-const { pollResponse } = require("@modules/polls");
+const { pollResponse } = require("@services/poll-service");
 const { httpPermCheck } = require("@modules/middleware/permission-check");
 const { parseJson } = require("@modules/middleware/parse-json");
 const { isAuthenticated } = require("@modules/middleware/authentication");
