+++ api/v1/controllers/room/leave.js
@@ -1,5 +1,5 @@
 const { httpPermCheck } = require("@modules/middleware/permission-check");
-const { leaveRoom } = require("@modules/class/class");
+const { leaveRoom } = require("@services/room-service");
 const { isAuthenticated } = require("@modules/middleware/authentication");
 
 module.exports = (router) => {
