+++ jsconfig.json
@@ -5,7 +5,8 @@
             "@modules/*": ["modules/*"],
             "@services/*": ["services/*"],
             "@controllers/*": ["api/v1/controllers/*"],
-            "@errors/*": ["errors/*"]
+            "@errors/*": ["errors/*"],
+            "@sockets/*": ["sockets/*"]
         }
     },
     "include": ["**/*.js"]
